+++ src/main/kotlin/com/deepromeet/atcha/notification/api/request/NotificationRequest.kt
@@ -1,4 +1,4 @@
-package com.deepromeet.atcha.notification.domatin
+package com.deepromeet.atcha.notification.api.request
 
 data class NotificationRequest(
     val lastRouteId: String
